Fix 'use after free' of `ProgressBar` when closing a mpris player
The `ProgressBar` is a child of a `MessageList.Message` (which descends
from `Clutter.Actor`). On the gnome-shell side, once the mpris player
`hide`s (which seems to emit before `closed`), the corresponding
`MessageList.Message` (that is the parent of `ProgressBar`!) is
destroyed. And when destroying a `Clutter.Actor`, all of its children
are destroyed too, resulting in the `ProgressBar` being destroyed before
the `closed` signal, and thus the 'use after free'. So, we can simply
remove the signal handler.